// apps/web/src/features/cluster/types/display-data.ts
/**
 * Represents the possible display data fields for a cluster card.
 */
export type ClusterCardDisplayData =
  | 'argocd'
  | 'grafana'
  | 'rorcli'
  | 'cpu'
  | 'memory'
  | 'nodes'
  | 'price'
  | 'agentVersion'
  | 'kubernetesVersion'
  | 'toolingVersion'
  | 'datacenterName'
  | 'datacenterProvider'
  | 'environment'
  | 'serviceId'
  | 'region'
  | 'az'
  | 'workspace'